split the line drawer
